fix: changes journalService to events from subjects
Once a journal event was detected and processed a subject would emit it. However this meant each subscriber had to have logic to test for type of event. Switching to events the journal service now emits events with the same name as those in the journal and so subscribers can simply listen for those cutting down on boiler plate code.